requirements.txt
django
django-admin-interface
django-crispy-forms
pillow
crispy-bootstrap4
requests